fix: UI bug in edit profile, where input label overlays on top of input value
Turns out, ReactHookForm really does not like undefined values.
Previously, the first(and default) option in the RHFSelect components (e.g. Pronouns selector, Highest Lead Rope Grade selector) was
undefined. This was not an issue during onboarding because the user does not have any profile information yet.

However, during Edit Profile, when we try to populate the input fields with existing data, the RHFSelect component will still think that there is no data and hence the input label will not shift up (until the user makes a change to the selection).

Hence, the fix is to simply change the first(and default) option to a blank string instead of undefined.