fix: Auto-filled jios clogging up the feed
Problem is that auto-filled jios have a start_date of when it is
created. This can be solved by reversing the sorting order. There
is no need to sort by start_date first, as the primary case for
sorting by start_date is to get an ascending order in time for jios.